rename : main news view model 위치 이동
